Made the mole up/down look smoother and changed the GUI to center the holes properly

+++ src/server/WAMGame.java
@@ -63,7 +63,7 @@ public class WAMGame implements Runnable{
      * @return Random Up Time
      */
     public long getRandomUptime(){
-        return wam.getRandomUpTime();
+        return (long)wam.getRandomUpTime();
     }
 
     /**
@@ -71,7 +71,7 @@ public class WAMGame implements Runnable{
      * @return Random Mole Time
      */
     public long getRandomDowntime(){
-        return wam.getRandomDownTime();
+        return (long)wam.getRandomDownTime();
     }
 
     /**
